pyproject.toml
[project]
name = "2025-fastapi-seminar-template-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.116.1",
    "freezegun>=1.5.5",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
]
